FIX: Move fireworks -> my_fireworks to avoid package conflict
